% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auctestr.R
\docType{package}
\name{auctestr}
\alias{auctestr}
\alias{auctestr-package}
\title{auctestr: Statistical Testing for AUC data.}
\description{
auctestr currently provides four main useful functions for statistical testing of the AUC, or A', statistic:
fbh_auc_compare, stouffer_z, fbh_test, and se_auc.
}
