% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbh_auc_compare.R
\name{se_auc}
\alias{se_auc}
\title{Compute standard error of AUC score, using its equivalence to the Wilcoxon statistic (See Fogarty, Baker and Hudson, "Case Studies in the use of ROC Curve Analysis for Sensor-Based Estimates in Human Computer Interaction" 2008.)}
\usage{
se_auc(auc, n_p, n_n)
}
\arguments{
\item{auc}{value of A' statistic (or AUC, or Area Under the Receiver operating characteristic curve) (numeric).}

\item{n_p}{number of positive cases (integer).}

\item{n_n}{number of negative cases (integer).}
}
\description{
Compute standard error of AUC score, using its equivalence to the Wilcoxon statistic (See Fogarty, Baker and Hudson, "Case Studies in the use of ROC Curve Analysis for Sensor-Based Estimates in Human Computer Interaction" 2008.)
}
