% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_compare.R
\name{stouffer_z}
\alias{stouffer_z}
\title{Compute aggregate z-score using Stouffer's method.}
\usage{
stouffer_z(z_vec, ignore.na = TRUE)
}
\arguments{
\item{z_vec}{vector of z-scores (numeric).}

\item{ignore.na}{should NA values be ignored? defaults to TRUE.}
}
\value{
numeric, Z-score using Stouffer's method aggregated over \code{z_vec}.
}
\description{
Compute aggregate z-score using Stouffer's method.
}
\references{
Stouffer, S.A.; Suchman, E.A.; DeVinney, L.C.; Star, S.A.;
Williams, R.M. Jr. The American Soldier, Vol.1: Adjustment during Army Life (1949).
}
