% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbh_auc_compare.R
\name{se_auc}
\alias{se_auc}
\title{Compute standard error of AUC score, using its equivalence to the Wilcoxon statistic.}
\usage{
se_auc(auc, n_p, n_n)
}
\arguments{
\item{auc}{value of A' statistic (or AUC, or Area Under the Receiver operating characteristic curve) (numeric).}

\item{n_p}{number of positive cases (integer).}

\item{n_n}{number of negative cases (integer).}
}
\description{
Compute standard error of AUC score, using its equivalence to the Wilcoxon statistic.
}
\seealso{
"The meaning and use of the area under a receiver operating characteristic (ROC) curve", Hanley and McNeil, Radiology, 1982.

Other fbh method: \code{\link{fbh_auc_compare}},
  \code{\link{fbh_test}}
}
